.pub-cache/hosted/pub.dartlang.org/html-0.15.0/test/data/tree-construction/isindex.dat

#data
<isindex>
#errors
#document
| <html>
|   <head>
|   <body>
|     <form>
|       <hr>
|       <label>
|         "This is a searchable index. Enter search keywords: "
|         <input>
|           name="isindex"
|       <hr>

#data
<isindex name="A" action="B" prompt="C" foo="D">
#errors
#document
| <html>
|   <head>
|   <body>
|     <form>
|       action="B"
|       <hr>
|       <label>
|         "C"
|         <input>
|           foo="D"
|           name="isindex"
|       <hr>

#data
<form><isindex>
#errors
#document
| <html>
|   <head>
|   <body>
|     <form>
